Moving CertificationsFeaturedCard to own file
